requirements.txt
# Basic requirements for TimeTracker application
tkinter
matplotlib
pillow  # For icon conversion